ref: object utility function wrapper
